requirements.txt
cffi==1.14.3
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
six==1.15.0
Werkzeug==1.0.1
flask-jsonpify==1.5.0
requests==2.25.0
python-docx==0.8.10
xlrd==1.2.0
pandas==1.1.4
openpyxl==3.0.5
XlsxWriter==1.3.7
